refactor: HttpResponse에 대한 빌더패턴을 HttpResponseBuilder로 이동
